Moving data helpers from another place
